requirements.txt
jupyter==1.1.1
python-dotenv==1.2.1
httpx==0.28.1
openai==2.8.1
tiktoken==0.12.0
tavily-python==0.7.13
langchain==1.0.8
langchain-core
langchain-openai==1.0.3
langchain-text-splitters==1.0.0
langchain-tavily==0.2.13
langchain-community==0.4.1
langchain-text-splitters==1.0.0

langgraph==1.0.3
langgraph-sdk==0.2.15
langgraph-cli[inmem]==0.4.11
langgraph-api==0.5.42
langgraph-checkpoint-sqlite==3.0.1
jsonschema-rs==0.29.1

fastapi[all]==0.121.3
uvicorn==0.38.0
setuptools==80.9.0
wheel==0.45.1

langserve==0.3.3